nearing completion on component transfer for setter form 2
