+++ react-app/src/components/Settings_Page/Setter_Form_2.js
@@ -1,6 +1,122 @@
-import React from 'react'
+import { useState } from 'react';
+import { useDispatch } from 'react-redux';
+import dropdownData from './dropdown_data.json';
+import { createUserSetting, updateThemeMedia } from '../../store/settings_store';
 
-export const SetterForm2 = () => {
+export const SetterForm2 = ({ style }) => {
+    const dispatch = useDispatch();
+
+    const fontSizesRaw = dropdownData['font-sizes'];
+    const fontFamiliesRaw = dropdownData['fonts'];
+    
+    const [accent_1, setAccent1] = useState(style.accent_1);
+    const [accent_2, setAccent2] = useState(style.accent_2);
+    const [accent_3, setAccent3] = useState(style.accent_3);
+    const [background_color, setBackgroundColor] = useState(style.background_color);
+    const [background_media_loading, setBackgroundMediaLoading] = useState(false);
+    const [background_media, setBackgroundMedia] = useState(style.background_media);
+    const [background_rotate, setBackgroundRotate] = useState(style.background_rotate);
+    const [font_family, setFontFamily] = useState(style.font_family);
+    const [font_size, setFontSize] = useState(style.font_size);
+    const [font_color, setFontColor] = useState(style.font_color);
+    const [p_f_2_btn, setPF2Btn] = useState('Edit');
+    const [p_f_2_disabled, setPF2Disabled] = useState(true);
+    const [theme_name, setThemeName] = useState('');
+    const [toggledState, toggleState] = useState(true);
+
+    const smpl = {
+        a_1: `${accent_1}`,
+        a_2: `${accent_2}`,
+        a_3: `${accent_3}`,
+        b_c: `${background_color}`,
+        f_c: `${font_color}`,
+        f_f: `${font_family}`,
+        f_s: `${font_size}`
+      }
+
+      const resetHandler = (e) => { 
+        const targID = e.target.dataset.settingId;
+        const targForm = document.getElementById(`sett-pg-setter-form-${targID}`);
+        const targFormKids = Array.from(targForm.children);
+
+        if (!toggledState) {
+            setPF2Disabled(true);
+            setPF2Btn('Edit');
+            toggleState(true);
+        }
+
+        targFormKids.forEach(targKid => {
+            switch (targKid.dataset.inputName) {
+                case 'Theme Name': setThemeName(style.theme_name || ''); break;
+                case 'Background Color': setBackgroundColor(style.background_color); break;
+                case 'Background Media': targKid.value = ''; break;
+                case 'Background Rotate': setBackgroundRotate(style.background_rotate); break;
+                case 'Font Color': setFontColor(style.font_color); break;
+                case 'Font Family': setFontFamily(style.font_family); break;
+                case 'Font Size': setFontSize(style.font_size); break;
+                case 'Accent 1': setAccent1(style.accent_1); break;
+                case 'Accent 2': setAccent2(style.accent_2); break;
+                case 'Accent 3': setAccent3(style.accent_3); break;
+                default: break;
+            }
+        })
+    }
+
+    const setBackgroundMediaHandler = (e) => {
+        const file = e.target.files[0];
+        if(file) setBackgroundMedia(file);
+    }
+
+    const createSettingHandler = async (e) => {
+        e.preventDefault();
+        
+        const targForm = e.target;
+        const targFormKids = Array.from(targForm.children);
+        const isSubmit = targFormKids.find(targKid => targKid.tagName === 'BUTTON' && targKid.innerText === 'Submit');
+        const hasMedia = targFormKids.find(targKid => targKid.dataset.inputName === 'Background Media');
+       
+        toggleState(prevState => !prevState);
+
+        if (!toggledState) {
+            setPF2Disabled(true)
+            setPF2Btn('Edit')
+        } else {
+            setPF2Disabled(false)
+            setPF2Btn('Submit')
+        }
+
+        if (p_f_2_btn === 'Submit') {
+            const user_id = user.id;
+    
+             await dispatch(createUserSetting({
+                user_id,
+                theme_name,
+                background_color,
+                background_rotate,
+                font_color,
+                font_family,
+                font_size,
+                accent_1,
+                accent_2,
+                accent_3,
+            }));
+            
+
+            if (isSubmit && hasMedia.value) { 
+                console.log('hit', e.target.dataset.settingId);
+
+                const formData = new FormData();
+                formData.append('media', background_media);
+                setBackgroundMediaLoading(true);
+                
+                dispatch(updateThemeMedia(user.id, formData));
+                setBackgroundMediaLoading(false);
+                window.location.reload();
+            }
+            resetHandler(e);
+        }
+    }
+
     return (
         <div 
             className='create-theme-container'
